requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
spacy
pandas
